requirements.txt
amqp==2.6.1
billiard==3.6.3.0
celery==4.4.7
Django==2.2.2
django-suit==0.2.28
django-templated-mail==1.1.1
djangorestframework==3.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djoser==2.0.5
kombu==4.6.11
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2020.1
redis==3.5.3
sqlparse==0.3.1
vine==1.3.0
